Add files via upload
The Mehandi Booking System is a digital platform that simplifies the process of booking professional Mehandi artists for weddings, festivals, and special events. It replaces the traditional, unorganized booking system with a structured, user-friendly interface for easy scheduling and selection. The system includes four modules: Book Now, Confirmation, Accessories, and Products, offering a seamless experience with real-time availability and additional services. Developed using HTML, CSS, Bootstrap, JavaScript, and Firebase, it ensures an interactive and responsive user experience. By integrating automation and e-commerce features, the platform enhances convenience for both customers and Mehandi professionals. It consists of 

1.	Book Now Module:
2.	Confirmation Module 
3.	Accessories Module 
4.	Products Module